move game creation tests to a separate file
